% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HuMMANet.R
\name{HuMMANetHub}
\alias{HuMMANetHub}
\title{HuMMANet ExperimentHub Records}
\usage{
HuMMANetHub(localHub = FALSE)
}
\arguments{
\item{localHub}{Logical scalar passed to `ExperimentHub::ExperimentHub()`.}
}
\value{
An `ExperimentHub` subset filtered to `package == "HuMMANet"`.
}
\description{
Return the `ExperimentHub` subset associated with `HuMMANet`.
}
\examples{
if (interactive()) {
  hub <- HuMMANetHub(localHub = TRUE)
  hub
}
}
